backend/pyproject.toml
[project]
name = "zwift-tv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "flask-sock>=0.7.0",
    "gevent>=25.9.1",
    "gunicorn>=23.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pre-commit>=4.4.0",
    "pytest>=9.0.1",
    "ruff>=0.14.5",
]

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
ignore = []

[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true
